Move file handling & mock server utils to test_utils crate
