+++ test_utils/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "test_utils"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+actix-web = {version = "4.3", features = ["openssl"]}
+base64 = "0.21"
+openssl = "0.10"
+serde_json = "1"
+tempfile = "3"
+unindent = "0.1"
